refactor: first part of a clang-tidy refactor

+++ packages/react-native-audio-api/common/cpp/audioapi/core/effects/DelayNode.cpp
@@ -30,7 +30,7 @@ void DelayNode::onInputDisabled() {
   numberOfEnabledInputNodes_ -= 1;
   if (isEnabled() && numberOfEnabledInputNodes_ == 0) {
     signalledToStop_ = true;
-    remainingFrames_ = delayTimeParam_->getValue() * getContextSampleRate();
+    remainingFrames_ = static_cast<int>(delayTimeParam_->getValue() * getContextSampleRate());
   }
 }
 
@@ -100,7 +100,8 @@ std::shared_ptr<AudioBuffer> DelayNode::processNode(
   }
 
   auto delayTime = delayTimeParam_->processKRateParam(framesToProcess, context->getCurrentTime());
-  size_t writeIndex = static_cast<size_t>(readIndex_ + delayTime * context->getSampleRate()) %
+  size_t writeIndex =
+      static_cast<size_t>(static_cast<float>(readIndex_) + delayTime * context->getSampleRate()) %
       delayBuffer_->getSize();
   delayBufferOperation(
       processingBuffer, framesToProcess, writeIndex, DelayNode::BufferAction::WRITE);
